Информация об изображениях перенесена из файла в БД

+++ StoreCMS/ImagesManagement/ImagesManagementFunctions/DeleteDependentImages.cs
@@ -9,15 +9,18 @@ namespace Treynessen.ImagesManagement
     {
         public static void DeleteDependentImages(string pathToImageFolder, string imageFullName)
         {
-            string imageName = imageFullName.Substring(0, imageFullName.LastIndexOf('.'));
-            string imageExtension = imageFullName.Substring(imageFullName.LastIndexOf('.'));
-            var images = Directory.GetFiles(pathToImageFolder, $"*{imageName}*{imageExtension}");
-            Regex regex = new Regex($"{imageName}(_\\d+x\\d+)?(_q\\d{{1,3}})?{imageExtension}$");
-            string pathToImage = $"{pathToImageFolder}{imageFullName}";
-            var forDelete = images.Where(i => !i.Equals(pathToImage, StringComparison.InvariantCulture) && regex.IsMatch(i)).ToList();
-            foreach (var f in forDelete)
+            if (Directory.Exists(pathToImageFolder))
             {
-                File.Delete(f);
+                string imageName = imageFullName.Substring(0, imageFullName.LastIndexOf('.'));
+                string imageExtension = imageFullName.Substring(imageFullName.LastIndexOf('.'));
+                var images = Directory.GetFiles(pathToImageFolder, $"*{imageName}*{imageExtension}");
+                Regex regex = new Regex($"{imageName}(_\\d+x\\d+)?(_q\\d{{1,3}})?{imageExtension}$");
+                string pathToImage = $"{pathToImageFolder}{imageFullName}";
+                var forDelete = images.Where(i => !i.Equals(pathToImage, StringComparison.InvariantCulture) && regex.IsMatch(i)).ToList();
+                foreach (var f in forDelete)
+                {
+                    File.Delete(f);
+                }
             }
         }
     }
